\name{GGS-package}
\alias{GGS-package}
\alias{GGS}
\docType{package}
\title{
\packageTitle{GGS}
}
\description{
\packageDescription{GGS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GGS}
\packageIndices{GGS}

This is the R port of the Python module GGS available on Github.  
}
\author{
\packageAuthor{GGS}

Maintainer: \packageMaintainer{GGS}
}
\references{
Hallac, Nystrup, and Boyd (2019). Greedy Gaussian Segmentation of Multivariate Time Series. 
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
